+++ src/julia-syntax.scm
@@ -153,11 +153,7 @@
 ;; except for rest arg
 (define (method-lambda-expr argl body rett)
   (let ((argl (map arg-name argl))
-        (body (if (and (pair? body) (eq? (car body) 'block))
-                  (if (null? (cdr body))
-                      `(block (null))
-                      body)
-                  `(block ,body))))
+        (body (blockify body)))
     `(lambda ,argl ()
              (scope-block
               ,(if (equal? rett '(core Any))
@@ -359,7 +355,7 @@
                                                            types)))
                                  (call (core svec) ,@temps)))
                           ,body ,isstaged))))
-       (if (symbol? name)
+       (if (or (symbol? name) (globalref? name))
            `(block (method ,name) ,mdef (unnecessary ,name))  ;; return the function
            mdef)))))
 
@@ -381,9 +377,7 @@
                        (if (nospecialize-meta? a) (caddr a) a))
                      kargl))
          (pargl (cdr argl))   ;; positional args
-         (body  (if (and (pair? body) (eq? (car body) 'block))
-                    body
-                    `(block ,body)))
+         (body  (blockify body))
          (ftype (decl-type (car pargl)))
          ;; 1-element list of vararg argument, or empty if none
          (vararg (let ((l (if (null? pargl) '() (last pargl))))
@@ -1011,7 +1005,7 @@
          (dcl   (and (pair? name) (eq? (car name) '|::|)))
          (rett  (if dcl (caddr name) '(core Any)))
          (name  (if dcl (cadr name) name)))
-    (cond ((and (length= e 2) (symbol? name))
+    (cond ((and (length= e 2) (or (symbol? name) (globalref? name)))
            (if (or (eq? name 'true) (eq? name 'false))
                (error (string "invalid function name \"" name "\"")))
            `(method ,name))
@@ -1681,19 +1675,18 @@
   ;; (for (= lhs X) body)
   (let ((coll  (make-ssavalue))
         (state (gensy)))
-    `(scope-block
-      (block (= ,coll ,(expand-forms X))
-             (= ,state (call (top start) ,coll))
-             ,(expand-forms
-               `(,while
-                 (call (top !) (call (top done) ,coll ,state))
-                 (scope-block
-                  (block
-                   ;; NOTE: enable this to force loop-local var
-                   #;,@(map (lambda (v) `(local ,v)) (lhs-vars lhs))
-                   ,(lower-tuple-assignment (list lhs state)
-                                            `(call (top next) ,coll ,state))
-                   ,body))))))))
+    `(block (= ,coll ,(expand-forms X))
+            (= ,state (call (top start) ,coll))
+            ,(expand-forms
+              `(,while
+                (call (top !) (call (top done) ,coll ,state))
+                (scope-block
+                 (block
+                  ;; NOTE: enable this to force loop-local var
+                  #;,@(map (lambda (v) `(local ,v)) (lhs-vars lhs))
+                  ,(lower-tuple-assignment (list lhs state)
+                                           `(call (top next) ,coll ,state))
+                  ,body)))))))
 
 ;; convert an operator parsed as (op a b) to (call op a b)
 (define (syntactic-op-to-call e)
@@ -2369,7 +2362,7 @@
                      ,.(apply append rows)))
             `(call (top typed_vcat) ,t ,@a)))))
 
-   '|'|  (lambda (e) `(call ctranspose ,(expand-forms (cadr e))))
+   '|'|  (lambda (e) `(call adjoint ,(expand-forms (cadr e))))
    '|.'| (lambda (e) `(call  transpose ,(expand-forms (cadr e))))
 
    'generator
@@ -2628,9 +2621,7 @@
                    (set-car! (cddr lam)
                              (append real-new-vars real-new-vars-def (caddr lam))))
                (insert-after-meta ;; return the new, expanded scope-block
-                (if (and (pair? body) (eq? (car body) 'block))
-                    body
-                    `(block ,body))
+                (blockify body)
                 (append! (map (lambda (v) `(local ,v)) real-new-vars)
                          (map (lambda (v) `(local-def ,v)) real-new-vars-def)))))
         ((eq? (car e) 'module)
@@ -3458,15 +3449,23 @@ f(x) = yt(x)
                   (else #f)))
           (case (car e)
             ((call new foreigncall)
-             (let* ((args (if (eq? (car e) 'foreigncall)
-                              ;; NOTE: 2nd to 5th arguments of ccall must be left in place
-                              ;;       the 1st should be compiled if an atom.
-                              (append (list)
-                                      (cond (atom? (cadr e) (compile-args (list (cadr e)) break-labels linearize-args))
-                                            (else (cadr e)))
-                                      (list-head (cddr e) 4)
-                                      (compile-args (list-tail e 6) break-labels linearize-args))
-                              (compile-args (cdr e) break-labels linearize-args)))
+             (let* ((args
+                     (cond ((eq? (car e) 'foreigncall)
+                            ;; NOTE: 2nd to 5th arguments of ccall must be left in place
+                            ;;       the 1st should be compiled if an atom.
+                            (append (if (atom? (cadr e))
+                                        (compile-args (list (cadr e)) break-labels linearize-args)
+                                        (list (cadr e)))
+                                    (list-head (cddr e) 4)
+                                    (compile-args (list-tail e 6) break-labels linearize-args)))
+                           ;; TODO: evaluate first argument to cglobal some other way
+                           ((and (length> e 2)
+                                 (or (eq? (cadr e) 'cglobal)
+                                     (equal? (cadr e) '(outerref cglobal))))
+                            (list* (cadr e) (caddr e)
+                                   (compile-args (cdddr e) break-labels linearize-args)))
+                           (else
+                            (compile-args (cdr e) break-labels linearize-args))))
                     (callex (cons (car e) args)))
                (cond (tail (emit-return callex))
                      (value callex)
@@ -3674,26 +3673,28 @@ f(x) = yt(x)
 
             ;; top level expressions returning values
             ((abstract_type primitive_type struct_type thunk toplevel module)
-             (case (car e)
-               ((abstract_type)
-                (let* ((para (compile (caddr e) break-labels #t #f))
-                       (supe (compile (cadddr e) break-labels #t #f)))
-                  (emit `(abstract_type ,(cadr e) ,para ,supe))))
-               ((primitive_type)
-                (let* ((para (compile (caddr e) break-labels #t #f))
-                       (supe (compile (list-ref e 4) break-labels #t #f)))
-                  (emit `(primitive_type ,(cadr e) ,para ,(cadddr e) ,supe))))
-               ((struct_type)
-                (let* ((para (compile (caddr e) break-labels #t #f))
-                       (supe (compile (list-ref e 4) break-labels #t #f))
-                       ;; struct_type has an unconventional evaluation rule that
-                       ;; needs to do work around the evaluation of the field types,
-                       ;; so the field type expressions need to be kept in place as
-                       ;; much as possible. (part of issue #21923)
-                       (ftys (compile (list-ref e 5) break-labels #t #f #f)))
-                  (emit `(struct_type ,(cadr e) ,para ,(cadddr e) ,supe ,ftys ,@(list-tail e 6)))))
-               (else
-                (emit e)))
+             (with-bindings
+              ((*very-linear-mode* #f))  ;; type defs use nonstandard evaluation order
+              (case (car e)
+                ((abstract_type)
+                 (let* ((para (compile (caddr e) break-labels #t #f))
+                        (supe (compile (cadddr e) break-labels #t #f)))
+                   (emit `(abstract_type ,(cadr e) ,para ,supe))))
+                ((primitive_type)
+                 (let* ((para (compile (caddr e) break-labels #t #f))
+                        (supe (compile (list-ref e 4) break-labels #t #f)))
+                   (emit `(primitive_type ,(cadr e) ,para ,(cadddr e) ,supe))))
+                ((struct_type)
+                 (let* ((para (compile (caddr e) break-labels #t #f))
+                        (supe (compile (list-ref e 4) break-labels #t #f))
+                        ;; struct_type has an unconventional evaluation rule that
+                        ;; needs to do work around the evaluation of the field types,
+                        ;; so the field type expressions need to be kept in place as
+                        ;; much as possible. (part of issue #21923)
+                        (ftys (compile (list-ref e 5) break-labels #t #f #f)))
+                   (emit `(struct_type ,(cadr e) ,para ,(cadddr e) ,supe ,ftys ,@(list-tail e 6)))))
+                (else
+                 (emit e))))
              (if tail (emit-return '(null)))
              '(null))
 
@@ -3758,12 +3759,23 @@ f(x) = yt(x)
                         '() arg-map))
           body))))
 
-;; find newvar nodes that are unnecessary because (1) the variable is not
+(define (for-each-isdefined f e)
+  (cond ((or (atom? e) (quoted? e)) #f)
+        ((and (pair? e) (eq? (car e) 'isdefined))
+         (f (cadr e)))
+        (else
+         (for-each (lambda (x) (for-each-isdefined f x))
+                   (cdr e)))))
+
+;; Find newvar nodes that are unnecessary because (1) the variable is not
 ;; captured, and (2) the variable is assigned before any branches.
-;; this is used to remove newvar nodes that are not needed for re-initializing
-;; variables to undefined (see issue #11065). it doesn't look for variable
-;; *uses*, because any variables used-before-def that also pass this test
-;; are *always* used undefined, and therefore don't need to be *re*-initialized.
+;; This is used to remove newvar nodes that are not needed for re-initializing
+;; variables to undefined (see issue #11065).
+;; It doesn't look for variable *uses*, because any variables used-before-def
+;; that also pass this test are *always* used undefined, and therefore don't need
+;; to be *re*-initialized.
+;; The one exception to that is `@isdefined`, which can observe an undefined
+;; variable without throwing an error.
 (define (definitely-initialized-vars stmts vi)
   (let ((vars (table))
         (di   (table)))
@@ -3772,6 +3784,8 @@ f(x) = yt(x)
           di
           (begin
             (let ((e (car stmts)))
+              (for-each-isdefined (lambda (x) (if (has? vars x) (del! vars x)))
+                                  e)
               (cond ((and (pair? e) (eq? (car e) 'newvar))
                      (let ((vinf (var-info-for (cadr e) vi)))
                        (if (and vinf (not (vinfo:capt vinf)))